feature:  save with silex
