+++ startup/scripts/beforescenesave.py
@@ -1,0 +1,2 @@
+from silex_client.action.action_query import ActionQuery
+ActionQuery('save').execute()
